HDDS-6684. Move unit tests out of integration
